refactored sidebar component
