+++ src/components/Dashboard/Sidebar/Sidebar.jsx
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types';
-import React, { useEffect, useState } from 'react';
-import { Container, Image, Nav, Navbar } from 'react-bootstrap';
+import React, { useState } from 'react';
+import { Container, Image, Navbar } from 'react-bootstrap';
 import { useNavigate } from 'react-router-dom';
 
 import { logout } from '@api';
@@ -8,63 +8,7 @@ import kidsFirstLogo from '@media/logo/LOGO-BYME.png';
 
 import styles from './Sidebar.module.css';
 import { SIDEBAR_DATA } from './sidebarData';
-
-const SidebarItemsCard = ({
-  title,
-  icon,
-  activeIcon,
-  hoverIcon,
-  path,
-  isActive,
-  onClick,
-}) => {
-  const [isHovered, setIsHovered] = useState(false);
-  const navigate = useNavigate();
-
-  useEffect(() => {
-    if (isActive) {
-      navigate(path);
-    }
-  }, [isActive, path, navigate]);
-
-  const handleClick = (e) => {
-    e.preventDefault();
-    onClick(title, path);
-  };
-
-  const handleKeyDown = (e) => {
-    if (e.key === 'Enter' || e.key === ' ') {
-      e.preventDefault();
-      handleClick(e);
-    }
-  };
-
-  return (
-    <Nav.Item
-      className={`${styles.nav} ${isActive ? styles.active : ''}`}
-      id={window.location.pathname === path ? 'active' : ''}
-      onMouseEnter={() => setIsHovered(true)}
-      onMouseLeave={() => setIsHovered(false)}
-      onClick={handleClick}
-      role="button"
-      tabIndex={0}
-      onKeyDown={handleKeyDown}
-    >
-      {isActive ? activeIcon : (isHovered ? hoverIcon : icon)}
-      <span className={styles.sidebarMenuItem}>{title}</span>
-    </Nav.Item>
-  );
-};
-
-SidebarItemsCard.propTypes = {
-  title: PropTypes.string.isRequired,
-  icon: PropTypes.node.isRequired,
-  activeIcon: PropTypes.node.isRequired,
-  hoverIcon: PropTypes.node.isRequired,
-  path: PropTypes.string.isRequired,
-  isActive: PropTypes.bool.isRequired,
-  onClick: PropTypes.func.isRequired,
-};
+import SidebarItemsCard from './SidebarItemsCard';
 
 const Sidebar = ({ onTitleChange }) => {
   const [activeLink, setActiveLink] = useState(null);
@@ -145,17 +89,6 @@ const Sidebar = ({ onTitleChange }) => {
       </div>
     </Navbar>
   );
-};
-
-SidebarItemsCard.propTypes = {
-  title: PropTypes.string.isRequired,
-  icon: PropTypes.node.isRequired,
-  activeIcon: PropTypes.node.isRequired,
-  hoverIcon: PropTypes.node.isRequired,
-  path: PropTypes.string,
-  alt: PropTypes.string,
-  onClick: PropTypes.func,
-  isActive: PropTypes.bool,
 };
 
 Sidebar.propTypes = {
